+++ el-generator-cmdline.h
@@ -63,7 +63,7 @@ struct gengetopt_args_info
   const char *binary_help; /**< @brief File is in binary format help description.  */
   int neo4j_flag;	/**< @brief Output the CSV Neo4J expects (default=off).  */
   const char *neo4j_help; /**< @brief Output the CSV Neo4J expects help description.  */
-  int NE_chunk_size_arg;	/**< @brief Number of edges to generate in a chunk. (default='1048576').  */
+  long NE_chunk_size_arg;	/**< @brief Number of edges to generate in a chunk. (default='1048576').  */
   char * NE_chunk_size_orig;	/**< @brief Number of edges to generate in a chunk. original value given at command line.  */
   const char *NE_chunk_size_help; /**< @brief Number of edges to generate in a chunk. help description.  */
   int verbose_arg;	/**< @brief Provide status updates via stdout. (default='1').  */
